requirements.txt
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
email_validator==2.2.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-api-core==2.29.0
google-auth==2.48.0
google-cloud-core==2.5.0
google-cloud-firestore==2.16.1
google-cloud-secret-manager==2.20.2
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.62.3
gunicorn==22.0.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pluggy==1.6.0
proto-plus==1.27.0
protobuf==4.25.8
psycopg2-binary==2.9.9
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pytest==8.3.2
python-dotenv==1.0.1
requests==2.32.5
rsa==4.9.1
SQLAlchemy==2.0.32
stripe==14.3.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.0.3
WTForms==3.2.1
